Determine which Eurostat datasets need to be downloaded

Takes the translated specification, extracts the Eurostat variable codes, and uses a dictionary to check from which Eurostat datasets they should be retrieved.

Usage

determine_datacodes(specification, dictionary = NULL)

Arguments

specification

A tibble or data.frame as returned by translate_variables.

dictionary

A tibble or data.frame storing the Eurostat variable code in column 'eurostat_code' and the Eurostat dataset id in column 'dataset_id' . When NULL, the default dictionary is used.

Value

Returns a list with two character vectors:

var.ids

contains the Eurostat variable ids

data.ids

contains the Eurostat dataset ids.