Load Locally

Loads data from a local directory

Usage

load_locally(to_obtain, inputdata_directory, quiet)

Arguments

to_obtain

A data.frame as returned by determine_variables().

inputdata_directory

A path to .rds, .csv, or .xlsx input files in which the data is stored. Can be NULL, in which case data is obtained via download if possible.

quiet

Logical with default = FALSE. Should messages be displayed? These messages are intended to give more information about the estimation and data retrieval process.

Value

Returns a list with two named elements: $df stores the downloaded data and $to_obtain the updated data.frame tracking which variables still need to be obtained.